Move marshalling API into submodule
